// types/organization.ts
export interface IOrganization {
  name: string;
  displayName?: string;
  description: string;
  roles: IRole[];
  members: number;
  createdBy: string;
}

export type IOrganizationPayload = Pick<
  IOrganization,
  "name" | "description" | "displayName"
>;

export interface IMembership {
  userId: {
    _id: string;
    username: string;
    name: string;
    surname: string;
  };
  roles: IRole[];
}

export interface IRole {
  _id: string;
  name: string;
  description: string;
}

export type IRolePayload = Pick<IRole, "name" | "description">;

export interface IOrganizationSearchFilters {
  name?: string;
  displayName?: string;
  nameOrDisplayName?: string;
}